+++ src/super/i-block/i-block.ts
@@ -20,7 +20,7 @@ import log, { LogMessageOptions } from 'core/log';
 import { deprecated } from 'core/functools/deprecation';
 import { EventEmitter2 as EventEmitter } from 'eventemitter2';
 
-import RenderInfo from 'super/i-block/modules/render-info';
+import InfoRender from 'super/i-block/modules/info-render';
 import config from 'config';
 
 import Async, {
@@ -168,7 +168,7 @@ export * from 'super/i-block/modules/module-loader';
 
 export * from 'super/i-block/modules/daemons';
 export * from 'super/i-block/modules/event-emitter';
-export * from 'super/i-block/modules/render-info';
+export * from 'super/i-block/modules/info-render';
 
 export * from 'super/i-block/modules/sync';
 export * from 'super/i-block/modules/async-render';
@@ -786,10 +786,10 @@ export default abstract class iBlock extends ComponentInterface {
 	@system({
 		atom: true,
 		unique: true,
-		init: (ctx) => new RenderInfo(ctx)
+		init: (ctx) => new InfoRender(ctx)
 	})
 
-	readonly renderInfo!: RenderInfo;
+	readonly infoRender!: InfoRender;
 
 	/**
 	 * API for analytic engines
@@ -2577,8 +2577,8 @@ export default abstract class iBlock extends ComponentInterface {
 	 * Initializes data collection
 	 */
 	@hook(['mounted', 'updated'])
-	protected initRenderInfo(): void {
-		this.renderInfo.initDataGathering();
+	protected initInfoRender(): void {
+		this.infoRender.initDataGathering();
 	}
 
 	protected override onCreatedHook(): void {